requirements.txt
open_clip_torch
invisible-watermark
scipy
numpy
matplotlib
opencv_python
torch
torchvision
torchaudio
tqdm
Pillow
h5py
albumentations
diffusers
pudb
imageio
imageio-ffmpeg
pytorch-lightning
torch-fidelity
PySoundFile
transformers
torchmetrics
kornia==0.6
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
streamlit-drawable-canvas==0.8